Note the use of `enquote`
I just spent a good bit of time trying to figure this out and figured others might want to benefit. Basically, I want to use NSE and have the `f` be input as either a string or function. I was crashing RStudio when I tried to do so using `quote` and then came across `enquote`.