translation of plyr::ddply
This statement "most of dplyr is equivalent to ddply()[...]" was slightly misleading for me. I took me a while to realise that ddply can also be replaced by do() in some cases. After a couple of months of using dplyr, I realised this today while converting a function that returns a data frame, calculated on a window (GDP deflator by country).

This iris example might be best placed in a another page on translating plyr statement into dplyr.  But this introduction vignette is probably the first place most people will look at.