Improved performance when working with large number of columns. closes #879
The essence of the problem was that we created a large number of vector visitors, each of
which holds a Vector (e.g. NumericVector). The order of destruction of the vector
of visitor pointers was inducing a penalty because of the way R_PreserveObject and
R_ReleaseObject is implemented in R.