descriptive error messages
With the input

```
dat %> rename(dScore = {dPointsEarned / dPointsPossible})
```

I got

```
Error: Arguments to rename must be unquoted variable names. Arguments dScore are not.
```

which confused me. 




I still don't think my rewrite is prefect because it's unclear whether "argument" refers to the LHS or RHS of the `=`. (i.e., do we mean ` {dPointsEarned / dPointsPossible}` was wrong? Or that `dScore` was wrong? `dScore` is the argument to `rename`, but ` {dPointsEarned / dPointsPossible}` is the "argument to that argument". And if I'm confused and not taking a long time to think it over, it won't be immediately obvious which is which.






It might also be good to somehow note that `rename` is a function name (¿at least on colour terminals?). "Arguments to rename" sounds different to "The arguments passed to the function `rename`".