Add indirection for src$con
This commit adds con_acquire and con_release interfaces
which dplyr can use when it needs to get a connection from
a src (instead of just src$con). This makes it possible to
support connection pools properly; without this change,
there is no way for dplyr to signal to a connection pool
that it is done with a connection.