Update progress.R
tick function aborts and throws an error if number of ticks is exceeded. For the user, it is not desired to abort a time consuming task just because the initialization of a user feedback was carried out wrong. Therefore, I propose to just throw a warning and proceed with the task even though the progress bar is not reliable anymore.