Check return type of predicate in tbl_if_vars
If a predicate function returns a logical vector instead of a scalar or even a non-logical value, for instance in `select_if(iris, ~c(is.numeric(.x),is.numeric(.x)))` or in `select_if(iris, ~mean(.x))`, the error messages are "more elements supplied than there are to replace" and "Only strings can be converted to symbols", which is unclear about what went wrong. 
Also fixed a typo in the first `abort` call of the function.