Preserve ordered factors in `compute_groups()`
When making the new keys, it checked if the key was a factor or not -- but not if the key was ordered. So for any factor key, it created a new, unordered key via `new_factor()`.

I added an additional check if the key is ordered -- if so, it should create a new _ordered_ key in its place.

Fixes #5455