Remove `na_matches = "error"` support
We now think this will be part of `enforce(key_validate(a, b, c))`, which will check for missing values in the keys and for uniqueness.

Additionally, this again limits `na_matches` to being strictly about the join process. `"error"` could be validated independently from the join, but the other options can't be.