Adding an hyphen in How tos (#6162)
Although the hyphen is optional, one can easily argue that 'How-tos' is preferable to 'How tos'. A non-native speaker may not understand that the expression refers to an object (a guide), so it may look up the definition of 'tos'. Using 'How-tos' helps in this sense, as it indicates that it is a single expression.