Implement `join_by()` and more flexible join API (#5910)
* Hack in `vec_matches()` for testing

* Add prototype of `join_by()`

* Support `join_by()` with varying conditions in all joins

* Add remote on the `vec_matches()` vctrs PR

* Add support for controlling multiple matches

* Support the `filter` option of `vec_matches()`

* Let `vec_matches()` be responsible for dropping `x` specific rows

* Use `remaining` to avoid a call to `vec_in()`

* First pass of `join_by()` documentation

* Add `new_join_by()` constructor

* Implement `keep = NULL`

This defaults to `keep = FALSE` on equi conditions, but `keep = TRUE` on non-equi conditions, as this is generally what you want. The information in columns involved in a non-equi condition rarely overlaps, so you almost never want to drop the keys of the RHS.

* Revise `join_by()` examples section

* Let `vec_matches()` drop unmatched `needles` in `semi_join()`

* Clean up function signature styling

* Implement `complete` for forcing unmatched rows to error

* Implement `unique` for enforcing uniqueness in `x` or `y`

* Abstract out `dplyr_matches()`

* Give `nest_join()` `multiple`, `complete` and `unique` arguments

* Remove `check_multiple()` since `multiple` passes straight through

* Use `dplyr_matches()` in `join_filter()`

* Rename `unique` to `check_duplicates`

* Rename `complete` to `check_unmatched`.

This is much less ambiguous, and parallels nicely to `check_duplicates`.

* Take a pass at updating the documentation

* Implement `as_join_by()` and `join_by_common()`

These replace `standardise_join_by()` and separate logic and error handling in a cleaner way

* Push `anti_join()` and `semi_join()` through `join_rows()`

* Push `nest_join()` through `join_rows()`

* Second pass through `join_by()` documentation

* Implement `na_matches = "error"`

If you want to catch missing keys in `x` as an issue, this is the only way to do it. Missing keys in `y` can be caught by setting `check_unmatched = "y"`, since they will be treated as unmatched values.

* NEWS bullet

* Don't add `()` to binary operators

* Add `join_by()` tests

* Test join column generation with non-equi conditions

* Catch `multiple` related error and warning

Copying `mutate_cols()` for the catching of warnings

* Add tests to `join_rows()`

* Test that empty `join_by()` generates a cross join

* No longer need `is_join_by()`

* Add tests to `test-join.r`

* Add a note about the column name restriction of `join_by()`

* Only `setdiff()` away the equi key columns

Non-equi key columns should be considered in the duplicate check when `keep = NULL`, since both sides are kept

* Report the names of the duplicated columns, not the positions

The positions might be different from what the user supplied, since we rearrange them beforehand when we drop any duplicated non-equi column names

* Allow explicit referencing with `x$` and `y$` and add join helpers

There are three helpers: `between()`, `within()`, and `overlaps()`. If people need to tweak them at all, then they should change to constructing them from the binary conditions directly.

* Remove `na_matches = "error"` support

We now think this will be part of `enforce(key_validate(a, b, c))`, which will check for missing values in the keys and for uniqueness.

Additionally, this again limits `na_matches` to being strictly about the join process. `"error"` could be validated independently from the join, but the other options can't be.

* Remove `check_duplicates`

This will also be handled by something like `enforce(key_validate())`.

Again, note that this is a check that can be handled independently of the join process, so removing it seems logical.

* Replace `check_unmatched` with `unmatched`

Further analysis shows that `check_unmatched` is really only useful in a few scenarios. In particular, it is only helpful to avoid the surprising case of dropping rows, which can only occur in a few of the join types.

* Always throw enhanced warning from `dplyr_matches()`

There shouldn't be any performance issues from doing this, unlike #5686 where the warning was rethrown per group, which could be slow.

Removing this allows the enhanced warnings to be shown in `reprex()`s, see #6005.

* Change the default of `multiple` to `NULL`

This uses `"warning"` for equi and rolling joins, and `"all"` for cross joins and when there are any non equi join conditions present

* Class join specific errors and warnings

* Update tests affected by change to `multiple` default

* Use `.named = NULL` now that r-lib/rlang#1223 is fixed

* Update snapshots for dev vctrs

* Add `join_by()` to `_pkgdown.yml`

* Adjust position of the `multiple` NEWS bullet

* Add a non-equi join example to the main join page

* Expand on `keep = NULL` behavior

* Mention each type of join that is possible through `join_by()`

* Various `join_by()` doc updates from code review

* Mention data.table in NEWS bullet

* Tweak `join_by()` named expression error

* Use a switch statement to clean up control flow

* Tweak error message

* Further explain what "supplying a single name" does

* Use `needles_arg` and `haystack_arg` to improve error message generation

* Streamline if/else conditions in `standardise_` helpers

* Mention the multiple matches warning in the join description

* Various doc changes to main join help page based on feedback

* Move dots of `nest_join()` to right after `name = NULL`

So we can at least force the new arguments to be named

* Add `na_matches` to `nest_join()`

* NEWS bullet for `na_matches` in `nest_join()`

* Tweak `keep = NULL` documentation

* Adapt to removal of `missing = "propagate"` in `vec_matches()`

* Update snapshot tests for latest rlang

* Update for latest vctrs. `missing` -> `incomplete`

* Update to new vctrs PR

* Mix in a partial evaluation approach to match arguments by name

* Remove `max()` and `min()` in favor of `preceding()` and `following()`

* Use new faster `multiple = "any"` argument for semi/anti join

This is generally faster than first/last, sometimes by a lot (i.e. in the case where there would generally be a ton of matches)

* Update remotes and version requirements for vctrs and rlang

* Update join related snapshot tests

* Update `incomplete` from `"match"` to `"compare"`

And add a test for this behavior

* Handle cross joins on the dplyr side

`condition = NULL` is no longer allowed on the vctrs side, because we can easily emulate a cross join by generating two vectors of `1L`s to match with, and this makes the vctrs implementation much simpler.

* Use latest CRAN versions of rlang and vctrs

* Use uppercase `J`

* Be overly defensive against typos

* `by` is a required argument for `join_filter()`

* Pass the error call through `check_na_matches()`

* Pass the error call through `check_unmatched()`

* Defensively pass the `error_call` through more places in `join_mutate()`

I don't think these can error, as the common type is handled earlier, but it doesn't hurt and is good to be consistent

* Pass the error call through `join_by()`

* Pass the error call through `join_by_common()`

* Officially add support for `multiple = "any"`

* Make the `join_by()` helper documentation more user friendly

* Optimize `join_by()` docs for pkgdown and fix some typos

* Add a comment about the signatures of the `binding_*()` functions

* Upload revdep results

* Bump to 99 dev version so revdeps can depend on it