Add a `.locale` argument to `arrange()` and implement `dplyr_locale()` (#6263)
* Import unexported `vec_order_base/radix()`

* Use `vec_order_base()` in `with_order()` and `group_by()` internals

This allows us to isolate changes in this PR to just `arrange()`, keeping other usage of `vec_order()` the same for now.

* Add a `.locale` argument to `arrange()`, and implement `dplyr_locale()`

* Give scoped `arrange()` variants a `.locale` argument

Even though these are superseded, this should help ease the transition a little, since without this argument it would be difficult to choose a different locale, and if stringi wasn't installed then you'd unconditionally get a warning you couldn't silence

* NEWS bullet

* Mark `dplyr_locale()` documentation as internal

With the intention being that it shouldn't show up on the pkgdown site reference page. We expect most people to get to this page through `arrange()`.

* Default `dplyr_locale()` to C

This is reproducible across all R sessions and OSes, and is much faster, which makes it a good default.

It will also make the default of `arrange()` continue to align with what `group_by() + summarize()` returns, since that will also unconditionally use the C locale.

* Avoid saying "case-sensitivity" because that is a little confusing

* Move conversion of `locale` to `chr_proxy_collate` into `arrange_rows()`

This is a technical detail of `vec_order_radix()` so it is better to keep the conversion close to where we actually call it. Should make calling `arrange_rows()` on its own a little easier too.