Eliminate `mask$full_data()` and `private$data`
Working on `mutate(.when =)` showed that having `$full_data()` was a bit dangerous because the original data might now always reflect the real state of the data mask, i.e. with `.when` the real data was the original data sliced to where `.when` was `TRUE`.

Removing `$full_data()` actually makes it very easy to remove `private$data` too, which seems like a nice simplification