Retain original names from `y` in `join_cols()` output
It is actually less confusing to set the `y_key` names to `names(vars$x$key)` because that is what you are trying to really accomplish anyways - standardize to a single set of key names. This allows us to retain the original `y_key` names for use in a common cast error message.