Add caching to `pick()`
We do this by evaluating it only once on the original data frame. This greatly simplifies the caching, because we can just chop the picked columns from the original data frame, store the chops, and retrieve each group's chop on demand.

This means that `pick()` doesn't "update" sequentially, but instead we get the invariant that the number of rows returned by `pick()` is the same as the number of rows in the group, which feels more important.