Add an alias for `.data` so `?dplyr::.data` works
This should avoid this CRAN check note on r-devel, but we can't seem to reproduce it

```
* checking for missing documentation entries ... WARNING
Undocumented code objects:
  '.data'
All user-level objects in a package should have documentation entries.
See chapter 'Writing R documentation files' in the 'Writing R
Extensions' manual.
```

Purposefully not adding a bullet for `.data` in the documentation page, as it isn't actually a helper that is "no longer for normal usage".

In the long term we should move this to its own help page, possibly with `.env`, or remove it entirely.